sorted tag tests, added tests for shortId, fixed urlToId
