+++ tests/integration/tag/publication-tag-put.test.js
@@ -95,86 +95,6 @@ const test = async app => {
     }
   )
 
-  await tap.test('Remove Tag from Publication', async () => {
-    // before:
-    const pubresbefore = await request(app)
-      .get(`/publications/${urlToId(publication.id)}`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type('application/ld+json')
-
-    await tap.equal(pubresbefore.status, 200)
-    const bodybefore = pubresbefore.body
-    await tap.ok(Array.isArray(bodybefore.tags))
-    await tap.equal(bodybefore.tags.length, 1)
-
-    const res = await request(app)
-      .delete(`/publications/${publicationId}/tags/${tagId}`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type('application/ld+json')
-
-    await tap.equal(res.status, 204)
-
-    // after:
-    const pubres = await request(app)
-      .get(`/publications/${publicationId}`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type('application/ld+json')
-
-    await tap.equal(pubres.status, 200)
-    const body = pubres.body
-    await tap.ok(Array.isArray(body.tags))
-    await tap.equal(body.tags.length, 0)
-  })
-
-  await tap.test(
-    'Try to remove a Tag from a Publication with invalid Tag',
-    async () => {
-      const res = await request(app)
-        .delete(`/publications/${publicationId}/tags/${invalidTagId}`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type('application/ld+json')
-
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(
-        error.message,
-        `Remove Tag from Publication Error: No Relation found between Tag ${invalidTagId} and Publication ${publicationId}`
-      )
-      await tap.equal(
-        error.details.requestUrl,
-        `/publications/${publicationId}/tags/${invalidTagId}`
-      )
-    }
-  )
-
-  await tap.test(
-    'Try to remove a Tag from a Publication with invalid Publication',
-    async () => {
-      const res = await request(app)
-        .delete(`/publications/${invalidPubId}/tags/${tagId}`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type('application/ld+json')
-
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(
-        error.message,
-        `Remove Tag from Publication Error: No Relation found between Tag ${tagId} and Publication ${invalidPubId}`
-      )
-      await tap.equal(
-        error.details.requestUrl,
-        `/publications/${invalidPubId}/tags/${tagId}`
-      )
-    }
-  )
-
   await tap.test('Try to assign Publication to Tag twice', async () => {
     await request(app)
       .put(`/publications/${publicationId}/tags/${tagId}`)
