+++ tests/integration/tag/tag-post.test.js
@@ -22,6 +22,8 @@ const test = async app => {
       )
 
     await tap.equal(res.status, 201)
+    await tap.type(res.body.id, 'string')
+    await tap.equal(res.body.shortId, urlToId(res.body.id))
     await tap.equal(res.body.tagType, 'reader:Stack')
     await tap.equal(res.body.name, 'mystack')
     await tap.equal(res.body.json.property, 'value')
